refactor(webapp): fold MonthStepper into a single MonthSelector
The month picker only ever renders in two places now — the paid usage
header and the Free caps header — and both want the same size, so the
separate MonthStepper wrapper and its size prop added nothing. Inline it
back into MonthSelector (keeping the shared useSelectedMonth hook, the
fixed-width label, and the focus ring) and drop MonthStepper.
